Add ability to present circular menu buttons in a counterclockwise layout
There is a use case in which presenting circular menu buttons in a counterclockwise fashion is very helpful, especially when the FloatingButton is located at the top of a view instead of the bottom of a view. This change reverses the path and demonstrates the change in the FloatingButtonExample code.